Bluetooth: Move HRS implementation to subsys/bluetooth/lib
This moves GATT HRS Service implementation under
subsys/bluetooth/lib and integrate with Kbuild to avoid build
problems.

Fixes #1166
