+++ samples/bluetooth/peripheral/prj.conf
@@ -10,3 +10,4 @@ CONFIG_BT_DEVICE_APPEARANCE=833
 CONFIG_BT_BAS=y
 CONFIG_BT_CTS=y
 CONFIG_BT_DIS=y
+CONFIG_BT_HRS=y
