+++ samples/bluetooth/peripheral/src/main.c
@@ -22,8 +22,7 @@
 #include <bluetooth/bas.h>
 #include <bluetooth/cts.h>
 #include <bluetooth/dis.h>
-
-#include <gatt/hrs.h>
+#include <bluetooth/hrs.h>
 
 #define DEVICE_NAME		CONFIG_BT_DEVICE_NAME
 #define DEVICE_NAME_LEN		(sizeof(DEVICE_NAME) - 1)
