+++ samples/bluetooth/peripheral_hr/prj.conf
@@ -6,3 +6,4 @@ CONFIG_BT_DEVICE_NAME="Zephyr Heartrate Sensor"
 CONFIG_BT_DEVICE_APPEARANCE=833
 CONFIG_BT_BAS=y
 CONFIG_BT_DIS=y
+CONFIG_BT_HRS=y
