+++ samples/bluetooth/peripheral_hr/src/Makefile
@@ -1,3 +1,1 @@
-ccflags-y +=-I${ZEPHYR_BASE}/samples/bluetooth
-
-obj-y = main.o hrs.o
+obj-y = main.o
