+++ subsys/bluetooth/lib/hrs.c
