+++ tests/bluetooth/shell/prj.conf
@@ -13,4 +13,5 @@ CONFIG_BT_TINYCRYPT_ECC=y
 CONFIG_CONSOLE_SHELL=y
 CONFIG_BT_SHELL=y
 CONFIG_BT_DEVICE_NAME="test shell"
+CONFIG_BT_HRS=y
 CONFIG_BT_L2CAP_TX_BUF_COUNT=6
