+++ tests/bluetooth/shell/prj_br.conf
@@ -15,3 +15,4 @@ CONFIG_BT_GATT_CLIENT=y
 CONFIG_BT_L2CAP_DYNAMIC_CHANNEL=y
 CONFIG_BT_TINYCRYPT_ECC=y
 CONFIG_BT_DEVICE_NAME="test shell"
+CONFIG_BT_HRS=y
