+++ tests/bluetooth/shell/src/Makefile
@@ -1,4 +1,3 @@
 ccflags-y +=-I$(ZEPHYR_BASE)/include/drivers
-ccflags-y +=-I$(ZEPHYR_BASE)/samples/bluetooth
 
-obj-y = main.o hrs.o
+obj-y = main.o
